README.md rework, siarg.h split, msvc support added
