+++ README.md
@@ -1,68 +1,59 @@
-# Sili
-![VM workflow Linux](https://github.com/EimaMei/sili-toolchain/actions/workflows/linux.yml/badge.svg)
-![VM workflow Windows](https://github.com/EimaMei/sili-toolchain/actions/workflows/windows.yml/badge.svg)
-![VM workflow MacOS](https://github.com/EimaMei/sili-toolchain/actions/workflows/macos.yml/badge.svg)
+Sili
+====
+<center>
+<br>
+	<img src=".github/lonicat-neris.png" alt="Lonicat" width="150">
+<br>
+<b><i>„Miau miau miau a miau'au miau a miau miau miau, ma miau miau miau miau miau miau miau miau miau miau miau miau miau miau miau miau miau miau miau'll miau miau miau a miau.“ - katė</i></b>
 
-Sili is a cross-platform standard library for modern C99 programming, providing
-many new features, improvements and fixes to the language's standard libraries,
-while still adhering to the main principles and appeals of C.
+---
 
-# Main goals
-## Modernize C
-The project aims to modernize C by implementing a proper standard library that is
-more feature-rich and understandable to read and write, while also containing a
-lot of quality of life changes for better readability and safety. It serves
-as a lighter alternative to the bloated C++ STL.
-## Performance & flexibility
-Sili strives to strike a balance between having great performance and flexibility
-that provide many options for developers when using the library.
-## Portability
-`sili.h`  is a single-header library, allowing for easy integration into any project.
-Furthermore, sili is designed to be cross-platform, supporting any OS that has a
-working C99 compiler.
+</center>
 
-# Some of the features present in the library
-- Architecture/OS/Compiler/Standard indentification macros (`SI_CPU_X86`,
-`SI_SYSTEM_WINDOWS`, `SI_COMPILER_CLANG`, `SI_STANDARD_C99`, `SI_STANDARD_ANSI`
-etc).
-- Useful macros (`typeof()`, `SI_ASSERT`/`SI_STATIC_ASSERT`, `SI_BIT`, `for_range`/`for_eachArr` etc).
-- Robust string implementation (`siString`, fully compatible with `char*`).
-- Better arrays (`siArray(<type>)`, compatible with other C arrays).
-- `char` specific functions (`si_charToLower`, `si_charIsSpace`, `si_hexDigitToInt` etc).
-- Hashtable functions (`si_hashtableMakeReserve`, `si_hashtableGetWithHash`, `si_hashtableSet`)
-- Path related functions (`si_pathExists`, `si_pathCopy`, `si_pathGetFullname` etc).
-- A cross-platform file IO functions (`si_fileOpen`, `si_fileReadContents`, `si_fileWriteLineAt` etc).
-- `siOptional` type similar to `std::optional`.
-- Cross-platform multi-thread support (`si_threadMake`, `si_threadRun`, `si_threadJoin`).
-- Easy to use benchmarking functions to check performance (`si_benchmarkRunsPerLoop`,
-`si_benchmarkExecutesPerMs`, `si_benchmarkLoopsAvg` etc).
-- An allocator system, making allocations 3x faster than the traditional `malloc`/`free`.
-- A general-use hash table.
-- Unicode functions (`si_utf8Decode`, `si_utf8ToUtf16String`, `si_utf16ToUtf8String`).
-- Bit manipulation funcionts (`si_numCountBitsU32`, `si_numRotateLeft`, `si_bytesToNumArr`,
-`SI_NUM_BIT_SET` etc).
-- DLL function loading (`si_dllLoad`, `si_dllProcAddress`).
-- Custom `printf` implementation, which includes C23 and new specifiers, as well
-as better cross-platform support (`%ll` and `%l` both equal to `i64`).
+<p align="center">
+Sili is a collection of cross-platform and single-header libraries for general
+programming in C99/C11. Many new fixes and exciting additions are made to the C 
+libraries for a more modern developer experience while still adhering to the main 
+principles and appeals of C: simplicity, performance and portability.
+</p>
+
+<p align="center">
+	<a href="https://github.com/EimaMei/sili-toolchain/actions/workflows/linux.yml"> 
+		<img src="https://github.com/EimaMei/sili-toolchain/actions/workflows/linux.yml/badge.svg">
+	</a>
+	<a href="https://github.com/EimaMei/sili-toolchain/actions/workflows/windows.yml"> 
+		<img src="https://github.com/EimaMei/sili-toolchain/actions/workflows/windows.yml/badge.svg">
+	</a>
+	<a href="https://github.com/EimaMei/sili-toolchain/actions/workflows/macos.yml"> 
+		<img src="https://github.com/EimaMei/sili-toolchain/actions/workflows/macos.yml/badge.svg">
+	</a>
+</p>
+
+# Libraries
+|  Name                | Version | Description |
+|----------------------|:-------:|:-----------:|
+|  [sili.h](sili.h)    | v0.1.0  | A general-purpose programming library to replace the C/C++ standard libraries.
+|  [sigar.h](sigar.h)  | v0.0.0  | A cross-platform library for audio playback and recording.
+|  [siarg.h](siarg.h)  | v0.0.0  | An argument-parsing library for making CLIs.
 
 # Examples
 Examples of most (but definitely not all) sili's features are utilized in the
 `examples` folder.
 
-# Release model
-Sili uses the standard major, minor and patch (e.g., `v1.2.5`) for its `main` 
-branch releases, which are meant to be stable and eligible for development. 
-They're guaranteed to pass all of the tests, have working examples and compiled
-binary releases available for Windows, MacOS and Linux.
+# Versioning systems
+## `main` branch
+- Format: `v<main>.<minor>.<patch>`.
+- Descrpition: A `main` release guarantees that all of the `sili` libraries pass
+their respective tests and have their examples working as intended, as well as 
+compiled releases being available for all platforms.   
 
-If you're using the `dev` branch of sili, then you must be aware that they follow 
-the versioning system of `v<current.main.version>-dev-<commit count for the release>`. 
-For example, if 3 commits have been released for the `dev` branch, while the `main`
-branch version is `v0.5.3`, then those are going to be referred as `v0.5.3-dev-1`, 
-`v0.5.3-dev-2` and `v0.5.3-dev-3` respectively. `dev` releases are less likely to
-have fully working examples or tests are compile and work as intended on every 
-platform. 
+## `dev` branch
+- Format: `v<main-branch-version>-dev-<commit-number>`.
+- Descrpition: A `dev` release is less likely to pass all of the tests and have
+working examples on every platform.  
 
-The `experimental` version of sili is not meant for any stable use and is reserved 
-for the most cutting-edge release of sili. The branch follows the versioning system
-of `v<current.main.version>-experimental-<commit ID>`.
+## `experimental` branch
+- Format: `v<main-branch-version>-experimental-<commit-ID>`.
+- Descrpition: An `experimental` release is not meant for any stable use and is 
+reserved  for the most cutting-edge release of sili. Errors and warnings are to 
+be expected.
